Case insensitive check for Bridge environment
Setting "BridgeEnv=DEV" environment variable does not work due to
case sensitive check[1]. This change will allow setting BridgeEnv
to "dev" or "DEV". The use case for this is to allow setting
"BridgeEnv=DEV" in BridgePF-infra as an EB env var and reuse the
BridgeEnv variable to set the bucket names to "DEV_CONSENTS_BUCKET".

[1] https://gist.github.com/anonymous/7f7cf706b0249f0fdb50ba696c18dcd2